WIP - Reorganized project structure - Renamed components folder's items - Added index.js and styles.scss to some components
