+++ client/src/components/Form/Input/index.js
@@ -1,0 +1,3 @@
+import Input from './Input';
+
+export default Input;
