+++ client/src/components/Form/RadioGroup/index.js
@@ -1,0 +1,3 @@
+import RadioGroup from './RadioGroup';
+
+export default RadioGroup;
