+++ client/src/components/Form/Select/index.js
@@ -1,0 +1,3 @@
+import Select from './Select';
+
+export default Select;
