+++ client/src/components/Form/index.js
@@ -1,0 +1,6 @@
+import Form from './Form';
+import Input from './Input';
+import RadioGroup from './RadioGroup/RadioGroup';
+import Select from './Select/Select';
+
+export default {Form, Input, RadioGroup, Select};
