+++ client/src/components/Modal/ConfirmModal/index.js
@@ -1,0 +1,3 @@
+import ConfirmModal from './ConfirmModal';
+
+export default ConfirmModal;
