+++ client/src/components/Modal/index.js
@@ -1,0 +1,4 @@
+import Modal from './Modal';
+import ConfirmModal from './ConfirmModal';
+
+export default {Modal, ConfirmModal};
