+++ client/src/components/Modal/modal.scss
