+++ client/src/components/Toast/ErrorToast/index.js
@@ -1,0 +1,3 @@
+import ErrorToast from './ErrorToast';
+
+export default ErrorToast;
