+++ client/src/components/Toast/SuccessToast/index.js
@@ -1,0 +1,3 @@
+import SuccessToast from './SuccessToast';
+
+export default SuccessToast;
