--- client/src/components/common/modal/modal.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import './modal.css';
-
-function Modal({ handleClose, show, children }) {
-    const showHideClassname = show ? "modal display-block" : "modal display-none";
-
-    return (
-        <div className={showHideClassname}>
-                {children}
-        </div>
-    );
-}
-
-export default Modal;
