--- client/src/components/tabs/acls.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Acls extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="Acls"/>
-                Acls
-            </div>
-        );
-    }
-}
-
-export default Acls;
