--- client/src/components/tabs/connect.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Connect extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="connect"/>
-                Connect
-            </div>
-        );
-    }
-}
-
-export default Connect;
