--- client/src/components/tabs/group.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Group extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="Consumer Groups"/>
-                Group
-            </div>
-        );
-    }
-}
-
-export default Group;
