--- client/src/components/tabs/node.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Node extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="Nodes"/>
-                Node
-            </div>
-        );
-    }
-}
-
-export default Node;
