--- client/src/components/tabs/schema.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Schema extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="Schema Registry"/>
-                Schema
-            </div>
-        );
-    }
-}
-
-export default Schema;
