--- client/src/components/tabs/tail.jsx
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-import Header from "../common/header";
-
-class Tail extends Component {
-    render() {
-        return (
-            <div id="content">
-                <Header title="Live Tail"/>
-                Tail
-            </div>
-        );
-    }
-}
-
-export default Tail;
