+++ client/src/container/tabs/acls.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Acls extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="Acls"/>
+                Acls
+            </div>
+        );
+    }
+}
+
+export default Acls;
