+++ client/src/container/tabs/connect.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Connect extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="connect"/>
+                Connect
+            </div>
+        );
+    }
+}
+
+export default Connect;
