+++ client/src/container/tabs/group.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Group extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="Consumer Groups"/>
+                Group
+            </div>
+        );
+    }
+}
+
+export default Group;
