+++ client/src/container/tabs/node.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Node extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="Nodes"/>
+                Node
+            </div>
+        );
+    }
+}
+
+export default Node;
