+++ client/src/container/tabs/schema.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Schema extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="Schema Registry"/>
+                Schema
+            </div>
+        );
+    }
+}
+
+export default Schema;
