+++ client/src/container/tabs/tail.jsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import Header from "../header";
+
+class Tail extends Component {
+    render() {
+        return (
+            <div id="content">
+                <Header title="Live Tail"/>
+                Tail
+            </div>
+        );
+    }
+}
+
+export default Tail;
